Moved OpenGL logic from NovelCamera to NovelRenderService.
