+++ CMakeLists.txt
@@ -89,7 +89,7 @@ set(NOVELRT_HEADERS
   src/RGBAConfig.h
   src/wrapper.hpp
   src/NovelCamera.h
-  src/GeoMatrix4.h)
+  src/GeoMatrix4.h src/CameraViewChangedEventArgs.h)
 
 set(NOVELRT_SOURCES
         src/NovelWorldObject.cpp
